+++ install.sh
@@ -1,6 +1,6 @@
 python setup.py sdist
 
-pip install dist/auto_kappa-0.3.tar.gz
+pip install dist/auto_kappa-0.4.0.tar.gz
 
 rm -r dist